Rearrange scanning related functions
As a preparation for scanning enhancements, move stuff around:
- to keep the size of task_database.c manageable, add the disc
  serial related functions to task_database_cue.c instead
- move manual scan functions to task_database.c, remove
  task_manual_content_scan.c
- align the dependent includes
